Changes to eliminate warnings at VC warning level 4
date	2000.03.03.18.08.00;	author rmoore1;	state Exp;

+++ source/components/executer/exoparg1.c
@@ -641,7 +641,7 @@ AmlExecMonadic2R (
 
         /* The object exists in the namespace, return TRUE */
 
-        RetDesc->Number.Value = -1;
+        RetDesc->Number.Value = (UINT32) -1;
         goto Cleanup;
         break;
